feat: added browserslist support
If a browserslists config is found in CWD or any parent folder we now check if every supported
browser requires a polyfill for a detected feature in the javascript bundle and if none do, we do
not add the feature to the polyfill.io url.

BREAKING CHANGE: If your project makes use of browserslist, you need to now check that you are happy
with your polyfill.io url being based on that config.

fix #10